+++ test/FunctionMetadataGeneratorTests/FunctionMetadataGeneratorTests.cs
@@ -6,10 +6,12 @@ using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Dynamic;
+using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Azure.Messaging.ServiceBus;
+using Azure.Storage.Blobs;
 using Microsoft.Azure.Functions.Tests;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
@@ -245,27 +247,60 @@ namespace Microsoft.Azure.Functions.SdkTests
         }
 
         [Fact]
-        public void StorageFunction_SDKTypeBindings()
+        public void BlobStorageFunctions_SDKTypeBindings()
         {
             var generator = new FunctionMetadataGenerator();
             var module = ModuleDefinition.ReadModule(_thisAssembly.Location);
-            var typeDef = TestUtility.GetTypeDefinition(typeof(SDKTypeBindings));
+            var typeDef = TestUtility.GetTypeDefinition(typeof(SDKTypeBindings_BlobStorage));
             var functions = generator.GenerateFunctionMetadata(typeDef);
             var extensions = generator.Extensions;
 
-            Assert.Equal(1, functions.Count());
+            Assert.Equal(5, functions.Count());
 
-            var blobToBlob = functions.Single(p => p.Name == "BlobToBlobFunction");
+            var blobStringToBlobStringFunction = functions.Single(p => p.Name == "BlobStringToBlobStringFunction");
 
-            ValidateFunction(blobToBlob, "BlobToBlobFunction", GetEntryPoint(nameof(SDKTypeBindings), nameof(SDKTypeBindings.BlobToBlob)),
+            ValidateFunction(blobStringToBlobStringFunction, "BlobStringToBlobStringFunction", GetEntryPoint(nameof(SDKTypeBindings_BlobStorage), nameof(SDKTypeBindings_BlobStorage.BlobStringToBlobStringFunction)),
                 b => ValidateBlobTrigger(b),
                 b => ValidateBlobInput(b),
                 b => ValidateBlobOutput(b));
 
+            
             AssertDictionary(extensions, new Dictionary<string, string>
             {
                 { "Microsoft.Azure.WebJobs.Extensions.Storage.Blobs", "5.1.0-beta.1" },
             });
+
+            var blobClientToBlobStringFunction = functions.Single(p => p.Name == "BlobClientToBlobStringFunction");
+
+            ValidateFunction(blobClientToBlobStringFunction, "BlobClientToBlobStringFunction", GetEntryPoint(nameof(SDKTypeBindings_BlobStorage), nameof(SDKTypeBindings_BlobStorage.BlobClientToBlobStreamFunction)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInput(b),
+                b => ValidateBlobOutput(b));
+
+
+            var blobUnsupportedTypeToBlobStringFunction = functions.Single(p => p.Name == "BlobUnsupportedTypeToBlobStringFunction");
+
+            ValidateFunction(blobUnsupportedTypeToBlobStringFunction, "BlobUnsupportedTypeToBlobStringFunction", GetEntryPoint(nameof(SDKTypeBindings_BlobStorage), nameof(SDKTypeBindings_BlobStorage.BlobUnsupportedTypeToBlobClientFunction)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInput(b),
+                b => ValidateBlobOutput(b));
+
+
+            var blobPocoToBlobUnsupportedType = functions.Single(p => p.Name == "BlobPocoToBlobUnsupportedType");
+
+            ValidateFunction(blobPocoToBlobUnsupportedType, "BlobPocoToBlobUnsupportedType", GetEntryPoint(nameof(SDKTypeBindings_BlobStorage), nameof(SDKTypeBindings_BlobStorage.BlobPocoToBlobUnsupportedType)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInput(b),
+                b => ValidateBlobOutput(b));
+
+
+            var blobByteArrayToBlobByteArrayFunction = functions.Single(p => p.Name == "BlobByteArrayToBlobByteArrayFunction");
+
+            ValidateFunction(blobByteArrayToBlobByteArrayFunction, "BlobByteArrayToBlobByteArrayFunction", GetEntryPoint(nameof(SDKTypeBindings_BlobStorage), nameof(SDKTypeBindings_BlobStorage.BlobByteArrayToBlobByteArrayFunction)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInput(b),
+                b => ValidateBlobOutput(b));
+
 
             void ValidateBlobTrigger(ExpandoObject b)
             {
@@ -307,36 +342,100 @@ namespace Microsoft.Azure.Functions.SdkTests
         }
 
         [Fact]
-        public void ServiceBusTriggerFunction_SDKTypeBindings()
+        public void BlobCollectionFunctions_SDKTypeBindings()
         {
             var generator = new FunctionMetadataGenerator();
             var module = ModuleDefinition.ReadModule(_thisAssembly.Location);
-            var typeDef = TestUtility.GetTypeDefinition(typeof(ServiceBusSDKBindings));
+            var typeDef = TestUtility.GetTypeDefinition(typeof(SDKTypeBindings_BlobCollection));
             var functions = generator.GenerateFunctionMetadata(typeDef);
             var extensions = generator.Extensions;
 
-            Assert.Equal(1, functions.Count());
-
-            var serviceBusTrigger = functions.Single(p => p.Name == "ServiceBusTriggerFunction");
-
-            ValidateFunction(serviceBusTrigger, "ServiceBusTriggerFunction",
-                GetEntryPoint(nameof(ServiceBusSDKBindings), nameof(ServiceBusSDKBindings.ServiceBusFunction)),
-                b => ValidateServiceBusTrigger(b));
+            Assert.Equal(4, functions.Count());
 
             AssertDictionary(extensions, new Dictionary<string, string>
             {
-                { "Microsoft.Azure.WebJobs.Extensions.ServiceBus", "5.9.0-alpha.20230203.1" },
+                { "Microsoft.Azure.WebJobs.Extensions.Storage.Blobs", "5.1.0-beta.1" },
             });
 
-            void ValidateServiceBusTrigger(ExpandoObject b)
+            var blobStringToBlobClientEnumerable = functions.Single(p => p.Name == "BlobStringToBlobClientEnumerable");
+
+            ValidateFunction(blobStringToBlobClientEnumerable, "BlobStringToBlobClientEnumerable", GetEntryPoint(nameof(SDKTypeBindings_BlobCollection), nameof(SDKTypeBindings_BlobCollection.BlobStringToBlobClientEnumerable)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInputForEnumerable(b),
+                b => ValidateBlobOutput(b));
+
+            void ValidateBlobInputForEnumerable(ExpandoObject b)
             {
                 AssertExpandoObject(b, new Dictionary<string, object>
                 {
-                    { "Name", "message" },
-                    { "Type", "serviceBusTrigger" },
+                    { "Name", "blobinput" },
+                    { "Type", "blob" },
                     { "Direction", "In" },
-                    { "queueName", "queue" },
-                    { "Properties", new Dictionary<String, Object>() { { "SupportsDeferredBinding", "True" } } }
+                    { "blobPath", "container2" },
+                    { "Cardinality", "Many" },
+                    { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding", "True" } } }
+                });
+            }
+
+            var blobStringToBlobStringArray = functions.Single(p => p.Name == "BlobStringToBlobStringArray");
+
+            ValidateFunction(blobStringToBlobStringArray, "BlobStringToBlobStringArray", GetEntryPoint(nameof(SDKTypeBindings_BlobCollection), nameof(SDKTypeBindings_BlobCollection.BlobStringToBlobStringArray)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInputForStringArray(b),
+                b => ValidateBlobOutput(b));
+
+            void ValidateBlobInputForStringArray(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "blobinput" },
+                    { "Type", "blob" },
+                    { "Direction", "In" },
+                    { "blobPath", "container2" },
+                    { "Cardinality", "Many" },
+                    { "DataType", "String" },
+                    { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding", "True" } } }
+                });
+            }
+
+            var blobStringToBlobPocoEnumerable = functions.Single(p => p.Name == "BlobStringToBlobPocoEnumerable");
+
+            ValidateFunction(blobStringToBlobPocoEnumerable, "BlobStringToBlobPocoEnumerable", GetEntryPoint(nameof(SDKTypeBindings_BlobCollection), nameof(SDKTypeBindings_BlobCollection.BlobStringToBlobPocoEnumerable)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInputForEnumerable(b),
+                b => ValidateBlobOutput(b));
+
+
+            var blobStringToBlobPocoArray = functions.Single(p => p.Name == "BlobStringToBlobPocoArray");
+
+            ValidateFunction(blobStringToBlobPocoArray, "BlobStringToBlobPocoArray", GetEntryPoint(nameof(SDKTypeBindings_BlobCollection), nameof(SDKTypeBindings_BlobCollection.BlobStringToBlobPocoArray)),
+                b => ValidateBlobTrigger(b),
+                b => ValidateBlobInputForEnumerable(b),
+                b => ValidateBlobOutput(b));
+
+
+            void ValidateBlobTrigger(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "blob" },
+                    { "Type", "blobTrigger" },
+                    { "Direction", "In" },
+                    { "path", "container2/%file%" },
+                    { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
+                });
+            }
+
+            void ValidateBlobOutput(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "$return" },
+                    { "Type", "blob" },
+                    { "Direction", "Out" },
+                    { "blobPath", "container1/hello.txt" },
+                    { "Connection", "MyOtherConnection" },
+                    { "Properties", new Dictionary<String, Object>() }
                 });
             }
         }
@@ -847,13 +946,90 @@ namespace Microsoft.Azure.Functions.SdkTests
             }
         }
 
-        private class SDKTypeBindings
+        private class SDKTypeBindings_BlobStorage
         {
-            [Function("BlobToBlobFunction")]
+            [Function("BlobStringToBlobStringFunction")]
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
-            public object BlobToBlob(
+            public object BlobStringToBlobStringFunction(
                 [BlobTrigger("container2/%file%")] string blob,
                 [BlobInput("container2/%file%")] string blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+
+            [Function("BlobClientToBlobStringFunction")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobClientToBlobStreamFunction(
+                [BlobTrigger("container2/%file%")] BlobClient blob,
+                [BlobInput("container2/%file%")] Stream blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function("BlobByteArrayToBlobByteArrayFunction")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobByteArrayToBlobByteArrayFunction(
+                [BlobTrigger("container2/%file%")] byte[] blob,
+                [BlobInput("container2/%file%")] byte[] blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function("BlobUnsupportedTypeToBlobStringFunction")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobUnsupportedTypeToBlobClientFunction(
+                [BlobTrigger("container2/%file%")] BinaryData blob,
+                [BlobInput("container2/%file%")] BlobClient blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function("BlobPocoToBlobUnsupportedType")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobPocoToBlobUnsupportedType(
+                [BlobTrigger("container2/%file%")] Poco blob,
+                [BlobInput("container2/%file%")] BinaryData blobinput)
+            {
+                throw new NotImplementedException();
+            }
+        }
+
+        private class SDKTypeBindings_BlobCollection
+        {
+            [Function("BlobStringToBlobStringArray")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobStringToBlobStringArray(
+                [BlobTrigger("container2/%file%")] string blob,
+                [BlobInput("container2", IsBatched = true)] string[] blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+
+            [Function("BlobStringToBlobClientEnumerable")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobStringToBlobClientEnumerable(
+                [BlobTrigger("container2/%file%")] string blob,
+                [BlobInput("container2", IsBatched = true)] IEnumerable<BlobClient> blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function("BlobStringToBlobPocoEnumerable")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobStringToBlobPocoEnumerable(
+                [BlobTrigger("container2/%file%")] string blob,
+                [BlobInput("container2", IsBatched = true)] IEnumerable<Poco> blobinput)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function("BlobStringToBlobPocoArray")]
+            [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
+            public object BlobStringToBlobPocoArray(
+                [BlobTrigger("container2/%file%")] string blob,
+                [BlobInput("container2", IsBatched = true)] Poco[] blobinput)
             {
                 throw new NotImplementedException();
             }