+++ test/SdkE2ETests/PublishTests.cs
@@ -75,8 +75,71 @@ namespace Microsoft.Azure.Functions.SdkE2ETests
             Assert.True(JToken.DeepEquals(extensionsJsonContents, expected), $"Actual: {extensionsJsonContents}{Environment.NewLine}Expected: {expected}");
 
             // Verify functions.metadata
-            TestUtility.ValidateFunctionsMetadata(functionsMetadataPath, "functions.metadata");
+            TestUtility.ValidateFunctionsMetadata(functionsMetadataPath, "Microsoft.Azure.Functions.SdkE2ETests.Contents.functions.metadata");
         }
+
+
+        [Fact]
+        public async Task Publish_SdkTypeBindings()
+        {
+            string outputDir = await TestUtility.InitializeTestAsync(_testOutputHelper, nameof(Publish_SdkTypeBindings));
+            await RunPublishTestForSdkTypeBindings(outputDir);
+        }
+
+        [Fact]
+        public async Task Publish_Rid_SdkTypeBindings()
+        {
+            string outputDir = await TestUtility.InitializeTestAsync(_testOutputHelper, nameof(Publish_Rid_SdkTypeBindings));
+            await RunPublishTestForSdkTypeBindings(outputDir, "-r win-x86");
+        }
+
+        private async Task RunPublishTestForSdkTypeBindings(string outputDir, string additionalParams = null)
+        {
+            // Name of the csproj
+            string projectFileDirectory = Path.Combine(TestUtility.SamplesRoot, "WorkerBindingSamples", "WorkerBindingSamples.csproj");
+
+            await TestUtility.RestoreAndPublishProjectAsync(projectFileDirectory, outputDir, additionalParams, _testOutputHelper);
+
+            // Make sure files are in /.azurefunctions
+            string azureFunctionsDir = Path.Combine(outputDir, ".azurefunctions");
+            Assert.True(Directory.Exists(azureFunctionsDir));
+            var files = Directory.EnumerateFiles(azureFunctionsDir);
+
+            // Verify files are present
+            string metadataLoaderPath = Path.Combine(azureFunctionsDir, "Microsoft.Azure.WebJobs.Extensions.FunctionMetadataLoader.dll");
+            string extensionsJsonPath = Path.Combine(outputDir, "extensions.json");
+            string functionsMetadataPath = Path.Combine(outputDir, "functions.metadata");
+            Assert.True(File.Exists(metadataLoaderPath));
+            Assert.True(File.Exists(extensionsJsonPath));
+            Assert.True(File.Exists(functionsMetadataPath));
+
+            // Verify extensions.json
+            JObject jObjects = JObject.Parse(File.ReadAllText(extensionsJsonPath));
+            JObject extensionsJsonContents = jObjects;
+            JToken expected = JObject.FromObject(new
+            {
+                extensions = new[]
+                {
+                    new Extension("CosmosDB",
+                        "Microsoft.Azure.WebJobs.Extensions.CosmosDB.CosmosDBWebJobsStartup, Microsoft.Azure.WebJobs.Extensions.CosmosDB, Version=4.2.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35",
+                        @"./.azurefunctions/Microsoft.Azure.WebJobs.Extensions.CosmosDB.dll"),
+                    new Extension("Startup",
+                        "Microsoft.Azure.WebJobs.Extensions.FunctionMetadataLoader.Startup, Microsoft.Azure.WebJobs.Extensions.FunctionMetadataLoader, Version=1.0.0.0, Culture=neutral, PublicKeyToken=551316b6919f366c",
+                        @"./.azurefunctions/Microsoft.Azure.WebJobs.Extensions.FunctionMetadataLoader.dll"),
+                    new Extension("AzureStorageBlobs",
+                        "Microsoft.Azure.WebJobs.Extensions.Storage.AzureStorageBlobsWebJobsStartup, Microsoft.Azure.WebJobs.Extensions.Storage.Blobs, Version=5.1.0.0, Culture=neutral, PublicKeyToken=92742159e12e44c8",
+                        @"./.azurefunctions/Microsoft.Azure.WebJobs.Extensions.Storage.Blobs.dll"),
+                    new Extension("AzureStorageQueues",
+                        "Microsoft.Azure.WebJobs.Extensions.Storage.AzureStorageQueuesWebJobsStartup, Microsoft.Azure.WebJobs.Extensions.Storage.Queues, Version=5.0.0.0, Culture=neutral, PublicKeyToken=92742159e12e44c8",
+                        @"./.azurefunctions/Microsoft.Azure.WebJobs.Extensions.Storage.Queues.dll")
+                }
+            });
+            Assert.True(JToken.DeepEquals(extensionsJsonContents, expected), $"Actual: {extensionsJsonContents}{Environment.NewLine}Expected: {expected}");
+
+            // Verify functions.metadata
+            TestUtility.ValidateFunctionsMetadata(functionsMetadataPath, "Microsoft.Azure.Functions.SdkE2ETests.Contents.WorkerBindingSamplesOutput.functions.metadata");
+        }
+
 
         private class Extension
         {